requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
dukpy==0.3.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
ipython==8.12.3
itsdangerous==2.2.0
javascripthon==0.12
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.2.2
pretty-errors==1.2.25
prettytable==3.10.0
prompt_toolkit==3.0.45
pure-eval==0.2.2
pycparser==2.22
pyecharts==2.0.5
pyecharts-javascripthon==0.0.6
Pygments==2.18.0
PyMySQL==1.1.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
yarg==0.1.9
zipp==3.19.1
